WIP adding behaviour to keeping squads in safe place

+++ crate/src/constants.rs
@@ -1,4 +1,4 @@
-pub const MATH_PI: f32 = std::f64::consts::PI as f32;
+pub const MATH_PI: f32 = std::f32::consts::PI;
 pub const MAX_NUMBER_ITEMS_IN_PRODUCTION_LINE: usize = 5;
 pub const NORMAL_SQUAD_RADIUS: f32 = 60.0;
 pub const MAX_SQUAD_SPREAD_FROM_CENTER_RADIUS: f32 = NORMAL_SQUAD_RADIUS + 100.0;
